Add files via upload
new

+++ ft_strlen.c
@@ -5,14 +5,16 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: hdargui <hdargui@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/10/22 11:25:46 by mguerraf          #+#    #+#             */
-/*   Updated: 2024/10/26 10:03:52 by hdargui          ###   ########.fr       */
+/*   Created: 2024/11/05 15:42:57 by hdargui           #+#    #+#             */
+/*   Updated: 2024/11/05 15:42:57 by hdargui          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
